tax rate is in BigDecimal now

+++ app/uk/gov/hmrc/plasticpackagingtaxreturns/controllers/ReturnsController.scala
@@ -291,7 +291,7 @@ class ReturnsController @Inject()(
 object ReturnsController {
 
   case class ReturnWithTaxRate(displayReturnJson: JsValue,
-                               taxRate: Double)
+                               taxRate: BigDecimal)
 
   object ReturnWithTaxRate {
     implicit val format: Writes[ReturnWithTaxRate] = Json.writes[ReturnWithTaxRate]
